Move setting from internal/ to pkg/
